structs merged to v2
